+++ src/services/qrCode.ts
